cambios compra

+++ src/main/java/co/edu/uniquindio/proyecto/dto/CompraGetDTO.java
@@ -18,7 +18,7 @@ public class CompraGetDTO {
 
     private List<DetalleCompraDTO> detalleCompraDTO;
 
-    private float valorTotal;
+    private Double valorTotal;
 
     private LocalDateTime fecha;
 }
